build(script): add log about moving libraries
This adds a log in the terminal while performing the build, to update the user on the last
destination of the Nx libraries. Contrary to the last log from ng-packagr saying that the build is
in `/dist`, it is moved right after in the `/buil`. To avoid the confusion, this log say the right
information.