fix(schematics): normalize path in findTargetProject
findTargetProject returns OS-specific paths which may not match the paths in projectNode.files. To
fix this, the path needs to be normalized the same way as it is done in
isRelativeImportIntoAnotherProject. Also, per default when an import points to a directory, "index"
should be added as a potential import file candidate.