fix(nx): rename browserlist file to browserslist (#1654)
According to [the Browserslist docs](https://github.com/browserslist/browserslist#queries),
the file should be named `browserslist`. The previous name had a missing
"s" in the middle.

I do not know how `browserslist` is integrated in Nx and the Angular
CLI, but it is possible that the configuration file was being ignored
entirely due to this naming error. When the `browserslist` CLI does not
find a configuration file, it uses a sensible default configuration, so
it would not be obvious that the configuration file was ignored.