fix(testing): handle cypress parameters when invoked through Angular CLI
When Cypress was invoked using `ng`, it  was throwing errors because no additional properties were allowed.
When allowed, all the extra options are passed as array in the `--` key, which is not currently handled.
This commit allows additional properties and makes sure that if any adittional options are passed they are parsed and incorporated correctly.

ISSUES CLOSED: #3571