feat(core): add option to skip serve step for cypress builder
The builder currently supports skipping the serve step by passing an empty value
for
devServerTarget. This is not intuitive and makes configuring this via the command line messy.
This
flag has the same behavior as passing an empty devServerTarget but is more user-friendly.
This
option can be useful for CI purposes where applications are previously built and want to use
the
same configuration for e2e testing but the serve step is unnecessary.

ISSUES CLOSED: #3744