fix(node): use webpack compilation hooks to emit package.json (#10229)
Use the webpack plugin system to create the package.json, this allows for developers to use plugins like
CleanWebpackPlugin which clears the dist before the compilation starts. If this plugin is used in
combination with the old method of generating the package.json, then it is deleted on compilation start.